InputManager.h/.cpp modified to allow for a second ability and revive buttons.
Particle.h/.cpp modified to allow for lifetime and velocity setting.

ParticleManager.h/.cpp modified so that SpawnFloatingText returns a pointer to the text itself.

Player.h/.cpp has had the basis of a generic ability system added to it.

Fixed an unsigned/signed mismatch in UnitManager.cpp

+++ src/UnitManager.cpp
@@ -64,7 +64,7 @@ bool UnitManager::Update(float a_fDeltaTime)
 	}
 
 	//Limit of units based on the current haunting level.
-	if(m_apkUnits.size() < ((GameScene*)SceneManager::GetCurrentScene())->GetGameInfo()->GetHauntingLevel() * 2)
+	if(m_apkUnits.size() < (unsigned int)((GameScene*)SceneManager::GetCurrentScene())->GetGameInfo()->GetHauntingLevel() * 2)
 	{
 		SpawnNewUnit();
 	}
